Align iOS to Android for codegen (#33806)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/33806

This Diff updates the iOS codegen, supporting a process that is the same as what Android does.

Before this PR, iOS was expecting a `package.json` with a `codegenConfig` of the following shape:
```
'codegenConfig': {
  'libraries': [
    <lib1>,
    <lib2>,
    ...
  ]
}
```

This setup is not Android friendly. The Android platform works without the `libraries` array and generating everything in a single point.

After this diff, also iOS works in a similar way, accepting also a `package.json` of the likes:

```
'codegenConfig': {
    'name': '<a name>',
    'type': 'all|components|module',
    'jsSrcsDir: '<path>'
}
```

The codegen will look into all the subfolders of the `jsSrcsDir` path, looking for `Native<module>.js` and `<custom>NativeComponents.js`

**About Backward Compatibility**: This first diff continue to support also the previous array-like version. It prints some deprecation message to inform the users about this change.

**Testing**: This diff introduces some unit tests for the `generate-artifacts-executor` which reads the `package.json`

## Changelog
[iOS][Added] - Add support for a single library settings in the CodeGen.

Reviewed By: cortinico

Differential Revision: D36314901

fbshipit-source-id: fa8fdeb0999587a337a02cf44ce6bd689401a11c