improving iOS solution announce screenreaderError
reapply the previous solution after doing a reset of the branch
changes deleted are saved in branch

https://github.com/facebook/react-native/compare/main...fabriziobertoglio1987:text-input-errors-set-text-and-selection?expand=1

Moving the head back to this commit
https://github.com/facebook/react-native/commit/29b99ba4de79e5d3d24f48571a8832022b0ecf95

Reapply the solution for iOS announcing multiple times the errorMessage
https://github.com/facebook/react-native/commit/e0bdf61c2d2b64a7d13fd1ea4088635834415054
https://github.com/facebook/react-native/commit/3467c5def6a85b9048e04bef4e04dee8e852023e

A similar problem was also detected on Android and solved with a similar
solution
https://github.com/fabriziobertoglio1987/react-native-notes/issues/12#issuecomment-1096086753

The reset was done to remove the changes to the implementation which
would use codegen.setTextAndSelection() instead of adding the prop to
c++

Explained in detail here
https://github.com/fabriziobertoglio1987/react-native-notes/issues/12#issuecomment-1132668333

we have 2 types of data flow in the TextInput

SCENARIO 1
Javascript  <--->  Native Android/iOS

JavaScript receives in callback like onChangeText updates on the value
of the text and selection position, then decides with an onLayoutEffect
callback to over-ride those updates with the value kept in internal
state. This is how JavaScript TextInput component works.

SCENARIO 2
CPP <----> Native Android/iOS

The state in CPP/Android/iOS keeps in sync.

In this case, we don't receive any native update from Android/iOS on the
value of the screenreaderError in Native side (not like for example the
text value which is passed back to JavaScript with the _onChange
callback).
https://github.com/fabriziobertoglio1987/react-native/blob/f2e23215ca14c3c630aa931cdd114187589ac0fb/Libraries/Components/TextInput/TextInput.js#L1115

For this reason, I don't think is a good idea including the
screenreaderError in the setTextAndSelection.