Cleanup Folders after codegen
Summary:
This diff adds a function to clean up folders and empty files (if any) after the codegen runs in the OSS.
To align how iOS and Android behave in the codegen, we now have to preemptively create folders that we may not need later.
Thanks to this function, we recursively the folders in the codegen and we delete them if they are empty.

## Changelog
[iOS][Added] - Add function to cleanup codegen folders

Differential Revision: D36594999

fbshipit-source-id: 1ed0fcc13b2d38ffcae22ed69774219333be030f