Move utilities out of `react_native_pods` - Part 2
Summary:
This Diff moves another part of the utilities from the `react_native_pods` file to a specific `utils.rb` file.

It adds tests for these utils and improve our test mocks.

The goal is to simplify the `react_native_pods.rb` so it's easier to work with it.

I decided to split this diff in 2 because it was becoming quite big.

## Changelog

[iOS][Changed] - Refactoring part of the react_native_pods.rb script

Differential Revision: D37006265

fbshipit-source-id: 580152f6f3bc97c78e23d1004695c60a718c875c