Move utilities out of r`eact_native_pods` - Part 2
Summary:
This Diff moves another part of the utilities from the `react_native_pods` file to a specific `utils.rb` file.

It adds tests for these utils and improve our test mocks.

The goal is to simplify the `react_native_pods.rb` so it's easier to work with it.

I decided to split this diff in 2 because it was becoming quite big.

## Changelog

[iOS][Changed] - Refactoring part of the react_native_pods.rb script

Differential Revision: D37006265

fbshipit-source-id: ebca9e5c4da7c4a594f240e10bb39e9ad87eea08