Install Fabric dependencies when the New Architecture is enabled
Summary:
While testing some recent changes on a nightly, I realized that if we install the dependencies using RCT_NEW_ARCH_ENABLED=1 and we forget to change the flag in the Podfile, Fabric dependencies won't be installed.

However, Fabric is required by the New Architecture. This fix makes sure that we install Fabric dependencies when the New Architecture is enabled.

## Changelog
[iOS][Changed] Install Fabric dependencies when RCT_NEW_ARCH_ENABLED=1

Differential Revision: D38786978

fbshipit-source-id: 86f0eee55d6d1ea48b3f786399c2ad8899cb6f88