updating examples in AccessibilityExample
This example feels a bit odd to me. The state itself is not concatenated into the announcement, which makes sense, as it would lead to potentially announcing conflicting state if two child elements had separate states like "checked" and "unchecked", but because of this I'm not sure what value this example is providing. If there was no state here and just the label it would work the exact same way.
I think we can probably just remove this one and maybe replace it with one where the parent element has a state and making sure that state is reflected even when the children are combined into it.
https://github.com/facebook/react-native/pull/33690#discussion_r939454427
https://github.com/facebook/react-native/pull/33690#discussion_r958336725

I don't think importantForAccessibility="yes" is necessary for any of these examples, so lets remove it on all of them just to make it clear that this isn't a prequalification for this feature working.
https://github.com/facebook/react-native/pull/33690#discussion_r939457620

Remove duplicate example (this is the same as the previous one)
https://github.com/facebook/react-native/pull/33690#discussion_r939454068