Fix FillRateHelper Accessing -1 Frame
Summary:
FillRateHelper accesses frame metrics based on passed state/cellsAroundViewport on scroll. cellsAroundViewport may be [0, -1] for no items. Usually this is captured by the check that item count is zero, but it is possible to have items without yet expanding state to render the items. This adds a check to bail early if we are not yet rendering any viewport-related cells.

Also renamed "state" to "cellsAroundViewport" for more consistent naming with the new list, where it is no longer the only state.

Reviewed By: javache

Differential Revision: D39345705

fbshipit-source-id: 198ab46ff2c8e8fe64076c9150edd4914dd745d7