Fix `LogBoxSymbolication-test.js`
Summary:
This test was causing uncaught promise rejections, because the mocked `symbolicateStackTrace` in this test returns a different type from the real one, causing `LogBoxSymbolication`'s unawaited, cached promises to reject.

This wasn't picked up as a test failure because our `promise` polyfill silently swallows uncaught rejections, but it does cause a failure if we use native promises in tests.

Changelog:
[Internal][Fixed] - Fix LogBoxSymbolication-test.js

Reviewed By: huntie

Differential Revision: D39418412

fbshipit-source-id: 0be8f1551c4a58dc47faf1597caf47271af40af2