WIP - column-reverse, handle short list, fix issues
Work in Progress
Summary:
- use column-reverse instead of reversing order of items
- handle short list (3 items) using column-reverse. see https://github.com/facebook/react-native/pull/34141#issuecomment-1242913564
- fix regression with the initial index position (inverted list needs to
  start from the last index instead of the first index)

More Info

"Adapt current solution to use column-reverse instead of position relative:
Remove logic that inverts the order of the list
Add column-reverse logic to style and contentContainerStyle
Test logic disabled when adding more items to FlatList"	Sep 15 2:01 PM

"Implement the solution in VirtualizedList
Apply column-reverse style depending on the contentHeight (scrollEnabled and talkbackStyle)
Review changes"	Sep 15 6:50 PM

"Review diff
Reintroduce example from commit
Test other scenarios (infinite list, scrollToTop, horizontal flatlist)"	Sep 19 10:59 AM

"Column-reverse introduced several regressions with infinite list and other scenarios:

Expected: https://user-images.githubusercontent.com/24992535/191270491-4abfc24d-84b7-4b44-93ca-9e9adee8b0df.png
Actual: https://user-images.githubusercontent.com/24992535/191270513-34573184-04c9-4f90-abef-9ea1fdf199df.png

The inverted flatlist does not start at the bottom
scrollTo a specific index works correctly with column-reverse:
Understand reason _lastOffsetFromBottomOfScreen is null. Debug _maybeCallOnEndReached which is responsible for setting the value of _lastOffsetFromBottomOfScreen. The value of lastOffset is used to scroll to the bottom, but the value of lastOffset is not set before calling _onContentSizeChanged
The previous issue is caused by triggering onEndReached when loading the FlatList. The FlatList should only render 10 items, but while adding the items to the FlatList, scroll to the top triggers to add 3 more items (11, 12, 13th).
Comment logic adding additional items when scrolling to the top in the inverted-flatlist example.
Fix the condition in the example and avoid adding new items when FlatList is loading.
The following regression was introduced after removing the logic to add more items to the flatlist onEndReached (infinite list). The FlatList displays a blank screen after loading for the first time. The issue is caused by column-reverse adding 1 additional page to the flatlist. The FlatList starts scrolled to the last page (an empty page), and the last page consists of an empty screen. After scrolling to the top, the BlankScreen disappears. The pages are not aligned correctly. The first page displays items 5 to 10; the second page displays an empty page. The correct behavior would be: - Page 1 => items 5 to 10 - Page 2 => Items 1 to 5 => Start Position here (inverted) The issue reproduces in the previous example after removing onLayout and onContentSizeChange.
The infinite list does not keep scroll position when scrolling up."	Sep 19 11:56 AM

"The FlatList displays a blank screen after loading for the first time. The issue is caused by not verifying that contentLength and visibleLength are not undefined.
Understand the reason onLayout and onContentSizeChange trigger the issue in the content position
Debug the value of contentLength and visibleLength"	Sep 19 4:11 PM

"Rewrite inversion style logic:
Add the logic from the flatlist inverted example to VirtList#scrollProps.
Avoid updates when contentLength and visibleLength are undefined."	Sep 20 12:49 PM

"Removing onLayout and onContentSizeChanged from the flatlist-inverted example triggers a regression. Items are displayed upside down. The inversionStyle scaleY: -1 is applied when TalkBack is enabled. The inversionStyle is applied on the Cell, but not the FlatList.
Verify that the inversionStyle applied on the Cell is causing the issue.
The issue was caused by adding inversionStyle to FlatList style when contentLength/visibileLength was undefined. Solved by adding a condition that changes style to flexDirection: “column”"	Sep 20 1:47 PM

"FlatList inverted does not start scrolled to the bottom (expected, actual) when TalkBack is enabled.
Find the reason scrollTo is not triggered. scrollToOffset is called onContentSizeChanged. The scrollTo is calculated using _lastOffsetFromBottomOfScreen, which is undefined. _lastOffsetFromBottomOfScreen is computed when the end of the page is reached (maybeCallOnEndReached).
Understand why maybeCallOnEndReached does not update _lastOffsetFromBottomOfScreen. The logic inside onContentSizeChanged is responsible for keeping the scroll position in an infinite list. The current behavior is correct; _lastOffsetFromBottomOfScreen is not updated because of the missing onEndReached callback.
Find logic responsible for triggering scrollTo bottom of the list after the first render and understand the root cause of the issue.
Review diff with the main branch.
_hasTriggeredInitialScrollToIndex is true. An initial scroll to an index was triggered with a different _scrollMetrics.contentLength. The FlatList scrolls down 1 page instead of 3 pages."	Sep 20 2:24 PM

"Implement the below solution to handle long lists that may trigger multiple onContentSizeChanged when loading for the first time. The current behavior is that the flatlist starts not scrolled to the end (for ex. Page 3/3), instead is scrolled at Page 2/3:
save the last _scrollMetrics.contentLength
save the last item in the list (the first one in an inverted list)
if the contentLength changes and the lastItem does not change, trigger the scroll to the bottom
Review existing solution - review the diff - try to prepend items
Adding logic to avoid issues with multiple scrollTo animations with TalkBack. onMomentumScrollEnd and other ScrollView APIs are not compatible with the TalkBack gestures "	Sep 20 9:12 PM