Fix resizeMode="cover" when image is resized (#1273)
* Fix resizeMode="cover" when image is resized

Since `RCTResizeModeCover` is implemented manually on macOS (iOS benefits from `UIViewContentModeScaleAspectFill` API), we need to reprocess the image any time the bounds of the view update. The `updateImage:` method was already called on resize, so this instead stores the original image in an ivar so it can call `RCTFillImagePreservingAspectRatio` from `updateImage:` instead.

* Fix rendering template images with tintColor

This fixes a regression introduced by the previous commit, where images would no longer render with the `tintColor` on release builds. We must call `setTemplate:` (using `.template` property is illegal in Objective-C++ since `template` is reserved in C++) on the newly resized image instead of the original image.

Confirmed icons look correct in the release build of Messenger Desktop.

Co-authored-by: Scott Kyle <skyle@fb.com>