Generate Custom Native State
Summary:
This diff introduces the generation of custom native states using basic types as we do with the Props.

To make it work, the custom types are already writte in the Props.h file, therefore the State.h file must import that other file to have access to the required types.

This diff adds and updates the tests for the State.

## Changelog
[General][Added] - Generate custom Native State

Differential Revision: D39816763

fbshipit-source-id: 7793b0cce27251fcbff57a452f132b587d40bbfd