Prototype - Storing the Hermes iOS tarballs on Maven.
Summary:
I'm not going to merge this as it is, this is mostly for discussions with the community for the sake of
RFC #508 https://github.com/react-native-community/discussions-and-proposals/pull/508

Here I'm setting up a :ReactAndroid:external-artifacts Gradle module which is responsible of
uploading arbitrary artifacts to Maven (Central or any other repo).

In this specific example I'm uploading the Hermes iOS tarball which is 500+ Mb.
In this module I've configured the auth with Sonatype for publishing and the GPG signing of the artifacts.

Changelog:
[Internal] [Changed] - Store the iOS Hermes tarball on Maven

Differential Revision: D39886167

fbshipit-source-id: 66edae7f83df35525dc8cccd68bb5bb3a5e02a12