Fix detectedCliPath_withCliFromNodeModules
Implementation now matches that of react.gradle

The test was also updated as it was previously passing when it should have been failing.

The original test scenario passed because it was doing the lookup from the directory above node_modules. This meant that, even though the nodeProcess lookup failed, the subsequent rootCliJs would work. This scenario now does a lookup from a subdirectory next to node_modules so that the rootCliJs lookup will fail if reached.