Circle CI: do not share .git cache across different operating systems
Summary:
We use workflows extensively in Circle CI, so caching a git checkout should help speed up tests when downstream jobs need to checkout the repository. In the current configuration, the Windows job is most likely to run and write to the .git cache first, which results in permission issues when the .git cache is loaded onto macOS or linux hosts.

By splitting the cache by architecture, we may lose on some reusability across jobs with distinct architectures, but it ensures we avoid cross-platform permission issues.

Changelog: [Internal]

Differential Revision: D40150781

fbshipit-source-id: 7d6412cb1c232910a5f2e4e049c2e06f7ee91630