Fix forward: Filtering platform in codegen
Summary:
This [commit](https://github.com/facebook/react-native/commit/7680bdeb4f96a8092393372a59c77a9d7b729cae) added the possibility to create a Codegen specs that are platform specific.
However, it also modifies how the codegen is invoked and we need to publish a new version of the `react-native-codegen` package on NPM before we can use that feature.

## CHANGELOG
[General][Fixed] - Remove usage of the codegen spec filtering until we publish a new version of the codegen.

Differential Revision: D40176447

fbshipit-source-id: c1e169e7f128aaf3d009ad8fb78a52718192a94d