Add ImageSource and ImageRequest to generate custom NativeState
Summary:
This Diff adds supports on the CustomState generation for ImageSource and ImageRequests, to enable NativeComponents to use the ImageSource loading provided by React Native.

To achieve this, I also had to fox some errors in the imports and to introduce some functions to decorate parameters.

This diff also introduces the tests for thise additional types in both generators and parsers.

## Changelog
[General][Added] - add support for ImageSource and ImageRequest in the State.

Differential Revision: https://internalfb.com/D39884889

fbshipit-source-id: c2921db68090ca506642f874b24128015811c7d7