Set isPressable for onPressIn/Out
Summary:
In react-native-windows, the ability to detect a press event on a virtual span is dependent on setting a native only `isPressable` prop.

This modifies the Text component to include `onPressIn` and `onPressOut`. As a side-effect, this will also lazily initialize the pressability config when `onPressIn` or `onPressOut` only is set.

Changelog:
[General][Fixed] - Pressability for text with only `onPressIn` / `onPressOut` props

Differential Revision: D40264432

fbshipit-source-id: 9c67bae02a4c3cb76d41e4acffe9363c8c309707