Setup gradle-nexus to automate closing and releasing staging repositories
Summary:
I'm setting up `gradle-nexus` on our build to simplify the closing and releasing of staging repositories.
As of now, you'll have to go to Sonatype UI and manually click the release button.

With this plugin, we can instruct CircleCI to do the publishing automatically for us as
part of the nightly/stable release process.

Changelog:
[Internal] [Changed] - Setup gradle-nexus to automate closing and releasing staging repositories

Differential Revision: D40342759

fbshipit-source-id: eb55b07ce1c01584f9e41f4840394d9581b9926a