Create the Parser interface
Summary:
This diff is the base to create a polimorphic behavior for TypeScript and Flow parser. This type will allow to share a lot of code between the parsers and also to keep their differences a part.

It will be the base diff/PR for further tasks in the Codegen umbrella Issue

## Changelog:
[General][Added] - Parser interface to divide parser logic.

Differential Revision: D40548707

fbshipit-source-id: 4f6b5e56d74b9b2d92b87dfcf36abab6cb222606