Add the `propConversion` header to generated props
Summary:
Generated props with custom structs generates implement their own version of `fromRawProps` in the header file, which has to invoke `fromRawProps` in the framework which is not available.

This Diff adds the missing header in the generated `Props.h` file and updates the tests.

This should also fix [this issue] https://github.com/facebook/react-native/issues/35054.

## Changelog:
[General][Fixed] - Add missing header in the Codegen for props with object types.

Differential Revision: D40674432

fbshipit-source-id: c18ea0941c252acfb9d42ef240940fccab2231a1