RNGP - Do the .so cleanup using pickFirst and exclude
Summary:
It turns out that using the Artifacts API to manipulate the APK to remove
.so has unintended side effects and is causing the `installDebug` and `installRelease`
commands to fail.

I've resorted to register a packaging option for each variant to make sure we include only
the correct artifacts we want.

This should fix the current startup crash that is experienced on main.

Changelog:
[Android] [Fixed] - RNGP - Do the .so cleanup using pickFirst and exclude

Differential Revision: D40722285

fbshipit-source-id: 9d933551cbc0ed39b753b84d3398e1fc8b7c6641