add span description
Notes
"The Screenreader stops reading the span text when the focus moves on the AccessibilityLink. The issue may be connected to the latest changes.
incompatible types: CharSequence cannot be converted to SpannableString spannableDescription = text.subSequence(start, end);
The start and end of the AccessibilityLink need to be updated after changing the length of the string. Complete the functionality and fix all the compile errors"	"https://github.com/facebook/react-native/pull/35130#issue-1428239936
https://github.com/fabriziobertoglio1987/react-native/blob/48a01d006a45c49018d8c3860f40b696b1aa7269/ReactAndroid/src/main/java/com/facebook/react/uimanager/ReactAccessibilityDelegate.java#L600"