Link against the app codegen if available
Summary:
This commit extends the ReactNative-application.cmake logic so
that if the app is using codegen at the app level, the generate library
is properly built and linked.

It helps us simplify the RN Tester setup and makes it easier for app users
to use the codegen at the app level.

Changelog:
[Internal] [Changed] - [Android] Link against the app codegen if available

Differential Revision: D40936941

fbshipit-source-id: 4ef9403ccc49a655dfbc8c60c72838bd857a0c3a