RNGP - Correctly Support Gradle Configuration Cache
Summary:
This little change allows to support Gradle Configuration cache in user projects:
https://docs.gradle.org/current/userguide/configuration_cache.html

It allows to save several seconds on the build time.
We'll keep it disabled for now, but Gradle plans to enable it by default for everyone
in the future, so this changes makes us ready for it.

Changelog:
[Internal] [Changed] - RNGP - Correctly Support Gradle Configuration Cache

Differential Revision: D41519506

fbshipit-source-id: 0d017da59a8cbf1419207ab529b79cce8f7e0453