Consolidate hermes-executor-debug and -release inside a single target
Summary:
Historically, we used to have hermes-executor debug and release as separate dynamic libraries.
This makes it impossible to prefab this library, so I have to reconcile it into a single library.

This will also help keep the setup consistent with the internal (BUCK) where we have a single target.

Changelog:
[Internal] [Changed] - Consolidate hermes-executor-debug and -release inside a single target

Differential Revision: D41519119

fbshipit-source-id: 3702d1011e46fcfdf0912d99935831789007a1b1