RNGP - Fix DependencyUtils for Windows support
Summary:
It turns out that my previous diff hasn't really solved the Windows support.
The problem is that we're not escaping the URI of the Maven Local repository
properly on Windows.

To overcome this, I'll instead use the `toURI()` Api of File to properly
create a valid URI for a given folder.

Changelog:
[Internal] [Changed] - RNGP - Fix DependencyUtils for Windows support

Differential Revision: D41581849

fbshipit-source-id: 489af2298a3eebf179ff76febe84030094841026