Avoid using shared_ptr when forwarding arguments in AsyncCallback
Summary:
[Changelog][Internal]

Removes usage of `shared_ptr` when forwarding arguments to an async callback invocation for C++ bridging.

The shared_ptr is not really required here, as we don't share the arguments beyond the current scope, so this is basically an incidental overhead/complexity.

Note that ideally we'd like to use e.g. `unique_ptr` instead, but since we are still using some older compilers, and [C++11 doesn't yet support capture of move-only types in lambdas](https://en.wikipedia.org/wiki/C%2B%2B14#Lambda_capture_expressions), I resorted to allocating the args tuple on the heap manually. I think this is OK in this case, since it's encapsulated/localized in one place (with one caveat that the code now implicitly assumes that the callback will be always called).

Differential Revision: D42265274

fbshipit-source-id: 5fd4b7d205ed046771812c6bb3814998c10c7ac8