Use lineHeight to position span textAlignVertical
The TextView lineHeight is set to the highest value. A span lineHeight may change other spans lineHeight. The textAlignment align correctly if lineHeight was set from another span.
TextAlignVertical does not align correctly when lineHeight is undefined.

Calculating the lineHeight as average between height and numberOfLines in the text may cause issue when one of the lines includes an InlineImage. The InlineImage changes the lineHeight of that line, but the other lines keep their default height. The average lineHeight changes for the other lines and it is not anymore an accurate estimate. The text is not correctly vertically aligned (Screenshot).

https://github.com/facebook/react-native/pull/35704#issuecomment-1372337476