Fix for resources not correctly bundlded on release appbundles
Summary:
When downgrading from AGP 7.4 to 7.3 we were forced to resort to older APIs
to bundle resources. It seems like we haven't properly wired the task to make sure
resources generated by Metro are correctly accounted before the generation of
release app bundles/apks.

This fixes it. This fix can also be removed once we are on AGP 7.4
Fixes #35865

Changelog:
[Android] [Fixed] - Fix for resources not correctly bundlded on release appbundles

Differential Revision: D42573450

fbshipit-source-id: f3f3ed228788e9c406e066b63e70f20fd0f4daeb