Reference implementation for NativePerformanceObserver
Summary:
[Changelog][Internal]

Add a minimal/reference JavaScript implementation for NativePerformanceObserver - the purpose is both unit testing (JS and native sides separately) and potentially shimming the part of functionality that is not dependent on native side.

This is both a setup for adding general unit tests for the Performance* APIs, but also to be able to do non-trivial changes on JS side for WebPerformance (such as in (D43154319).

Differential Revision: D43167392

fbshipit-source-id: 0b319171590fbe09384f05ad3d507b6869de204c