Implement durationThreshold option for PerformanceObserver
Summary:
[Changelog][Internal]

By [the W3C standard](https://developer.mozilla.org/en-US/docs/Web/API/PerformanceObserver/observe), `PerformanceObserver.observer` can optionally take a `durationThreshold` option, so that only entries with duration larger than the threshold are reported.

This diff adds support for this on the RN side.

NOTE: The standard suggests that default value for this is 104s. I left it at 0 for now, as for the RN use cases t may be to too high (needs discussion).

Differential Revision: https://internalfb.com/D43154319

fbshipit-source-id: 884ca22b62daf241d1b98facc58958dab1797633