Rename the `React-RCTFabric` module from `React` to `RCTFabric` to avoid name clashes
Summary:
Change the `React-RCTFabric` prefix and `module_name` from React to `RCTFabric`.

In OSS, React-Core and React-RCTFabric are two different podspec, which results in two different frameworks. We can't have two frameworks with the same `module_name` as this is a naming clash, therefore I had to rename the `React-RCTFabric` `module_name` from `React` to `CTFabric.

This change affected also Codegen which needed to be updated to reflect the new framework namespace.

## Changelog:
[iOS][Changed] - Change module_name and prefix of `React-RCTFabric` from `React` to `RCTFabric`

Differential Revision: https://internalfb.com/D43088702

fbshipit-source-id: a3253a9c75179e05068cb18ac1cbfc659024c8f2