Update ruby scripts to install the dependencies with `use_frameworks`.
Summary:
This diff update the Cocoapods scripts to install the proper dependencies and search paths when `use_frameworks!` is declared together with the New Architecture.

Practically, it adds the right search paths to the codegen and the project files and makes sure that third party dependencies that leverage the `install_modules_dependencies` are populated with the right search paths for Frameworks.

It also adds unit tests for the changes and the new methods introduced.

## Changelog:
[iOS][Changed] - Properly install dependencies with `use_frameworks!`

Differential Revision: https://internalfb.com/D43089869

fbshipit-source-id: 97c49d2047eac5a30493f01e31cf3b34b47019fa