not adding test for `IncorrectModuleRegistryCallTypeParameterParserError` in `parseModuleName` fn
- reason: The problem here is that;

```flow
export default TurboModuleRegistry.getEnforcing('SampleTurboModule');
```

this result in an `UntypedModuleRegistryCallParserError`, because it is "not generic".

The problem is that, to get a type for it, it "must" be a generic.
I can't think of any other way for `getEnforcing` to be typed with a different type rather than generic. 🤔

I don't think there is a way to create a valid Flow code which is typed with something that is not a generic.
I looked at the base test for that throw and they are manually creating an AST that satisfy that requirement,
but that AST probably can't be actually generated. 🤔

-- @cipolleschi