Remove the enableSeparateBuildPerCPUArchitecture from the template entirely (#36341)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36341

See https://github.com/react-native-community/discussions-and-proposals/issues/602 for more context.
TL;DR: this flag is used for an edge case. We should not expose it to every users but move it to a guide.

I'll publish a guide on the website on how to achieve the same feature.

Changelog:
[Android] [Changed] - Remove the enableSeparateBuildPerCPUArchitecture from the template entirely

Reviewed By: cipolleschi

Differential Revision: D43695574

fbshipit-source-id: 0a8bc7a169804de41e0b0cf813d184301635b907