Minimize EditText Spans 7/N: Avoid temp list
Summary:
This is part of a series of changes to minimize the number of spans committed to EditText, as a mitigation for platform issues on Samsung devices. See this [GitHub thread]( https://github.com/facebook/react-native/issues/35936#issuecomment-1411437789) for greater context on the platform behavior.

This change addresses some minor CR feedback and removes the temporary list of spans in favor of applying them directly.

Differential Revision: D44295190

fbshipit-source-id: 27ad202c799e7ebeb0baad1276bbde1442d8de04