Minimize EditText Spans 6/N: letterSpacing
Summary:
This is part of a series of changes to minimize the number of spans committed to EditText, as a mitigation for platform issues on Samsung devices. See this [GitHub thread]( https://github.com/facebook/react-native/issues/35936#issuecomment-1411437789) for greater context on the platform behavior.

This change lets us set `letterSpacing` on the EditText instead of using our custom span.

Changelog:
[Android][Fixed] - Minimize EditText Spans 6/N: letterSpacing

Differential Revision: https://internalfb.com/D44240777

fbshipit-source-id: 53ecfa3f46df695ad8917099528d61efbafccf5a