trigger codegenned reflection if module is a turbomodule
Summary:
Changelog: [Internal]

if we look at usages of ReactModuleWithSpec, we see that it's existence was to simply identify generated native modules since the inheritance graph of native modules was different if they were generated.

this was introduced before we created the TurboModule interface, which all codegenned native modules also conform to. since that exists now, there's no need for both of these. this is the only callsite in our code where ReactModuleWithSpec triggers any logic, so i'm updating it.

Differential Revision: D44450687

fbshipit-source-id: 622007e0bceb9d4984bde561f6dec240872dd1b8