add unit test for method reflection on generated native modules (#36671)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36671

Changelog: [Internal]

BaseJavaModuleTest tests the reflection of the native module methods wrapped by the JavaModuleWrapper abstraction, this abstraction is used by JNI in the C++ layer.

however, though we already have logic that supports the reflection of our codegenned modules (turbomodules), we didn't have a test for it yet, so i'm adding it in this change.

Differential Revision: D44450433

fbshipit-source-id: e704bda05202def3e1d5619fd81438bf0fef1031