Attach Bridgeless nativeModuleProxy in TurboModuleBinding (#36626)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36626

In Bridgeless mode, With the TurboModule interop layer, the TurboModule system will need to customize the nativeModuleProxy global.

This customization would be much easier if the nativeModuleProxy global were installed by the TurboModule system (and not the Bridgeless core).

So, this diff moves nativeModuleProxy installation into the TurboModule system.

Changelog: [Internal]

Reviewed By: cortinico

Differential Revision: D43993197

fbshipit-source-id: 035a66d272a773ff9382e7fc3dc543154b04d3c4