Attach Bridgeless nativeModuleProxy in TurboModuleBinding
Summary:
In Bridgeless mode, With the TurboModule interop layer, the TurboModule system will need to customize the nativeModuleProxy global.

This customization would be much easier if the nativeModuleProxy global were installed by the TurboModule system (and not the Bridgeless core).

So, this diff moves nativeModuleProxy installation into the TurboModule system.

Changelog: [Internal]

Differential Revision: https://internalfb.com/D43993197

fbshipit-source-id: 898851eff800f5151bb09c6e40a8e2f5cdeb7ba1