Modify AnimatedProps and AnimatedStyle to use AnimatedObject (#36741)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36741

AnimatedObject is a more generic version of AnimatedTransform, able to handle animated values within arrays and objects. This is useful for props of native components that may need to be animated per field.

This diff hooks up AnimatedObject to AnimatedProps and AnimatedStyle for values that are arrays or objects.

Changelog:
[Internal][Added] - Modify AnimatedProps and AnimatedStyle to use AnimatedObject

Reviewed By: rshest

Differential Revision: D44315336

fbshipit-source-id: 8b047a53971c60736d11ced46108b3cecfd41c6c