Introduce ObjectAnimatedNode (#36742)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36742

AnimatedObject is a more generic version of AnimatedTransform, able to handle animated values within arrays and objects. This is useful for props of native components that may need to be animated per field.

This diff adds the native (Android) counterpart to AnimatedObject node in JS. The node handles array and map value types.

Changelog:
[Internal][Added] - Introduce ObjectAnimatedNode Java-side node for handling array and object prop values

Reviewed By: mdvacca

Differential Revision: D44466563

fbshipit-source-id: 5c4326509a5a5e561ce9ceb2a7ca1d67be8aaee3