Add support to generate basic union types in TurboModules
Summary:
This diff is a first take at generating union types in TurboModules. Due to platform limitations (neither Android nor iOS have a union type in the language) we have to fallback to dynamic types in the base implementation.
The TM author has to manually handle the cast to those types.

Note that a similar mechanism is already in place for MixedType. With these changes we are, at least, bringing some type safety to JS.

## Changelog:
[General][Added] - Add generation of union types in TM

Differential Revision: https://internalfb.com/D44712170

fbshipit-source-id: d3eef207eb22e386904137d49efbe504bf8f54e7