Add support to generate basic union types in TurboModules (#36896)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36896

This diff is a first take at generating union types in TurboModules. Due to platform limitations (neither Android nor iOS have a union type in the language) we have to fallback to dynamic types in the base implementation.
The TM author has to manually handle the cast to those types.

Note that a similar mechanism is already in place for MixedType. With these changes we are, at least, bringing some type safety to JS.

## Changelog:
[General][Added] - Add generation of union types in TM

Differential Revision: D44712170

fbshipit-source-id: df710c97723e9c81bf3b6be37e3cf928838b5cfd