Implement scaffolding for the Pointer Capture imperative host ref APIs
Summary:
Changelog: [Internal] - Implement scaffodling for the Pointer Capture imperative host ref APIs on iOS

This diff is fundamentally about adding 3 new methods to `RCTSurfacePointerHandler` (`hasPointerCapture`, `setPointerCapture`, and `releasePointerCapture`), and make it possible to call those methods synchronously from `ReactFabricHostComponent.js`. This diff doesn't actually start the logic implementation of these methods as I specifically would like to get focused feedback on how I accomplished hooking this up.

The delegate tree I had to traverse is fairly long and involves a lot of files so to make this easier to review here's a little diagram I made to make heads and tails of this:

{F946626014}

For the most part the code is just traversing this class tree except for when we reach the `RCTSurfacePresenter` where we retrieve the `RCTFabricSurface` related to the given `ShadowView` by leveraging it's `surfaceId`. That surface contains touch/pointer handler we need to call but we cannot do so yet because we haven't resolved the `UIView` that corresponds to that `ShadowView` yet so we call into `RCTMountingManager` is able to resolve the view, and pass it the surface so that it can call the necessary methods.

Additionally — because I'm adding new methods on `SchedulerDelegate` the Android renderer has to update it's interface to have those methods in the `Binding` class but because I'm focused on iOS I've left them as no-ops to be implemented later.

Differential Revision: D44891230

fbshipit-source-id: d52dbd3796120beab55a9fb10bebeadb2520c822