Add example on how to use union types on TM
Summary:
This change is an alternative to show how to use union types in TM **without generating union types** natively.

The goal is to enable using unions in the JS layer, without compromising the Native typesafety.

If this approach is a good middle-ground, we can remove the MixedTypeAnnotation and migrate to this approach.

## Changelog:
[General][Added] - Add example on Union types for Turbomodules

Differential Revision: D45045552

fbshipit-source-id: 101e7036d83485e3b31802c2156d1e6dd6af9a64