make getModuleInstanceFromClass: required (#37009)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37009

Changelog: [iOS][Breaking]

continuing getting rid of optional methods, but this does have a behavioral change. if the delegate does not provide the module instance, then we lazily create it. before we were returning nil, idk why. this will be a breaking change if you have any classes that conform to `RCTTurboModuleManagerDelegate`

#saynotoruntimechecks

Reviewed By: cipolleschi

Differential Revision: D45022139

fbshipit-source-id: ca3a7ea9dd95c4a04755a5b164b5f05ae1aa18c5