Restore server controlled form fields to whatever they should be (#26708)
Summary:
Fizz can emit whatever it wants for the SSR version of these fields when
it's a function action so they might not align with what is in the
previous props. Therefore we need to force them to update if we're
updating to a non-function where they might be relevant again.

DiffTrain build for commit https://github.com/facebook/react/commit/2fa632381839c8732dad9107b90911163b7f2b7a.

Changelog: [Internal]

Reviewed By: poteto

Differential Revision: D45238365

fbshipit-source-id: b75b4779bd3759362d69f96994ba781982f650a0