Avoid retaining TurboModuleManager in AppDelegate (#37104)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37104

The AppDelegate (and other similar classes) only need to construct the TurboModuleManager for intialization purposes, and should not retain it beyond that. TurboModuleManager retains each of the TurboModule instances and through the new binding mechanism, also JSI pointers, which are invalid beyond the lifetime of the JS context.

Changelog: [iOS] Changed AppDelegate template to avoid retaining TurboModuleManager.

Reviewed By: sammy-SC, cipolleschi

Differential Revision: D45310066

fbshipit-source-id: 5263280e14d959f0d7530571e2614a59c8838088