W3CPointerEvents: fix pointercancel test scrollview on Android (#36959)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36959

Changelog: [Internal] - W3CPointerEvents: fix pointercancel test scrollview on Android

On Android the contents of the scrollview were not large enough to allow for scrolling (i.e. they didn't overflow the scroll container). This change sets the height of the content and container explicitly to ensure that the view can be scrolled.

Reviewed By: vincentriemer

Differential Revision: D45064333

fbshipit-source-id: 6aa264d569639bb7d906a3afd887985caa563070