Remove "auto" from YGFloatOptional conversion (#37186)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37186

"auto" is not a valid value to go into anything accepting `YGFloatOptional`. Remove the code here which special cases it (though we will produce the same result). Also remove error logging which is not localized/useful, in the theme of having user style inputs be handled gracefully.

Changelog: [Internal]

Reviewed By: yungsters

Differential Revision: D45434650

fbshipit-source-id: 5d28ff4005f9e28fb5d40eca4a83179dca8473dc