rename handleException to handleTaskFault (#37205)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37205

Changelog: [Internal]

so when i was trying to help mdvacca with figuring out what this API does, it was really confusing to grep for since `handleException` is a very common pattern in general. when i looked more closely at the usage, it was specifically used to propagate faulted venice tasks. so i'm updating the name here to be more explicit and help us with refactoring later.

Reviewed By: mdvacca

Differential Revision: D45501548

fbshipit-source-id: 1063148f799218bad86b62545690ea3b74800fc2