Pass the hermestag file to the `read` function (#37231)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37231

In a previous diff, we tried to disambiguate between versions of hermes engine by reading the hermestag file.

However, we forgot to pass the file to the `read` function.

It used to work both on main and on CI because that file doesn't exists in those environments. But it failed as soon as we tried to test a release.

## Changelog:
[iOS][Fixed] - Pass the hermestag file to the `read` function.

Reviewed By: dmytrorykun

Differential Revision: D45530010

fbshipit-source-id: c47be6a22c5b4436379ab5501654e43653e53859