Convert RootViewTest to Kotlin (#37227)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37227

This is a reference PR on how to convert a JVM Unit test from Java to Kotlin. A couple of things to keep in mind when converting:

1. Do not use hungarian notation (`mActivity` -> `activity`)
2. JUnit rules need to be annotated with `get:Rule` rather than just `Rule`.
3. Use `import org.powermock.api.mockito.PowerMockito.`when` as whenever`to avoid having to escape all the `when` function invocation in the code as it's a keyword.
4. Do static imports of all the Mockito/PowerMock functions.

Getting rid of PowerMock and using Fakes is a plus, but that's not always possible. Having the test running and be green is already a good result.

Changelog:
[Internal] [Changed] - Convert RootViewTest to Kotlin

Reviewed By: cipolleschi

Differential Revision: D45526517

fbshipit-source-id: 25545883eaa88cae5f6603670767fcca059a82a4