rename handleException to handleTaskFault (#37205)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37205

Changelog: [Internal]

so when i was trying to help mdvacca with figuring out what this API does, it was really confusing to grep for since `handleException` is a very common pattern in general. so i'm updating the name here to be more explicit and help us with refactoring later.

bypass-github-export-checks

Reviewed By: mdvacca

Differential Revision: D45501548

fbshipit-source-id: 6ae32328e1117ae81ba53f1742440a68aec38803