Reduce visibility of DevInternalSettings class (#37256)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37256

This class is public but annotated with VisibleForTesting which is not correct (as there are no tests for it).
Apparently this class is used only within this package so it's safe to reduce its visibility to package internal.

On top of this, we had methods which were not annotated with Override, which I've fixed + I've sorted the methods to have first all the public/Override ones and the then others after.

Changelog:
[Android] [Removed] - Reduce visibility of DevInternalSettings class

Reviewed By: mdvacca

Differential Revision: D45566473

fbshipit-source-id: 2c0e809094bb965f2d6263e7b2eae17a1f095165