DevServerHelper should not depend on internal ctor parameter (#37265)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37265

DevServerHelper was having a constructor parameter as `DevInternalSettings` which is effectively internal. This should not be the case as that class is Internal as was bleeding out of the public API.

I've updated the primary constructor to take instead:

```
  public DevServerHelper(
          DeveloperSettings settings,
          String packageName,
          InspectorPackagerConnection.BundleStatusProvider bundleStatusProvider,
          PackagerConnectionSettings packagerConnectionSettings) {
```

This is breaking change for users that were depending on the Internal class.

Changelog:
[Android] [Removed] - DevServerHelper should not depend on internal ctor parameter

Reviewed By: mdvacca

Differential Revision: D45600283

fbshipit-source-id: e73139dbdf5f2505201b2d2c8b5a9143b7e207ba