Fix emission of pending events in RN Android
Summary:
This diff removes an assertion that was causing a crash in production.
The root cause of the bug is that there's a race condition on viewState.mEventEmitter, the value of this field changes while the method is being executed.
In this diff I'm avoiding the crash by dispatching the event as soon as possible and unblock the fix in production
I creacted the task T152630743 to follow up in the root cause of the threading bug.

changelog: [internal] internal

Differential Revision: D45622656

fbshipit-source-id: 14d709f616666a7465a9b0cfd32e026f9899864f