delete RCTJSScriptLoaderModule (#37284)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37284

Changelog: [iOS][Removed]

in the past, we had this protocol so we could decorate the `RCTDevSplitBundleLoader`, which was removed by motiz88. now that doesn't exist anymore, we should delete this protocol. i did a quick grep on github and this protocol was only used by `RCTDevSplitBundleLoader`, so i don't think this will be a major breaking change

Reviewed By: christophpurrer

Differential Revision: D45591882

fbshipit-source-id: a366e445d246110f0dea76095ec0cf2682cbad0e