Fix emission of pending events in RN Android (#37290)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37290

This diff removes an assertion that was causing a crash in production.
The root cause of the bug is that there's a race condition on viewState.mEventEmitter, the value of this field changes while the method is being executed.
In this diff I'm avoiding the crash by dispatching the event as soon as possible and unblock the fix in production

changelog: [internal] internal

Reviewed By: fkgozali

Differential Revision: D45622656

fbshipit-source-id: 51caa0f3a2001fb33ed0efe0929ad112081f110e