Add @react-native/typescript-config
Summary:
This moves from `tsconfig/react-native`, a centralized repository of tsconfigs, to `react-native/typescript-config`, a package maintained as part of the RN monorepo.

We end up wanting to make changes where the versions are coupled, so this publishes them as part of the same repo, etc. It also means Meta engineers can more freely make changes to it with normal approval processes.

Differential Revision: https://internalfb.com/D45721088

fbshipit-source-id: 82704bebb9aaf1cc289f80ad2d92a6cdc2518609