Add `hasErrata()`, `addErrata()`, `removeErrata()`
Summary:
Adds internal helpers to YGConfig to make bit manipulation more readable. We also expose `hasErrata()` to YGNode beacuse checking that will be a common pattern. We intentionally don't add mutating functions to the node, since current model is to inval a node on commiting whole config.

This is not exposed via the C ABI.

Differential Revision: D45765971

fbshipit-source-id: f8e6ad737d5693cf2036b73bd7a5fa30fc0419d4