Add errata "PercentAbsoluteOmitsPadding" (#37376)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37376

X-link: https://github.com/facebook/yoga/pull/1290

This change default-enables a bugfix to Yoga's handling of absolute-positioned children sized using percentages. Yoga previously measured percentage of the box without padding, where the W3C spec, and browser behavior verified by test, use the padding box instead. See https://github.com/facebook/yoga/issues/850 for more details of the backing issue.

I have so far left this behind a `YGExperimentalFeature`, because it will change the dimensions of existing views which use absolute percentages with parent padding.

This change moves it from a `YGExperimentalFeature` to `YGErrata`. This means the conformant path is enabled by default, but users of `YGErrataClassic` and `YGErrataAll` will stay on the previous behavior.

Right now the name is `PercentAbsoluteOmitsPadding` as something that tries to explain the issue in lay-terms and be a little less long than the old one, but I am open to other ideas.

Reviewed By: yungsters

Differential Revision: D45763574

fbshipit-source-id: a5e17e2efd5478a236cf585f2bf67c6b823539e3