Add `hasErrata()`, `addErrata()`, `removeErrata()` (#37375)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37375

Adds internal helpers to YGConfig to make bit manipulation more readable. We also expose `hasErrata()` to YGNode beacuse checking that will be a common pattern. We intentionally don't add mutating functions to the node, since current model is to inval a node on commiting whole config.

This is not exposed via the C ABI.

Reviewed By: yungsters

Differential Revision: D45765971

fbshipit-source-id: e15fc66b38b96bad73f22f5530abdceb6b8ee44c