Android fno-rtti (#37380)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37380

We did the work to support this internally, let's see the savings we get with this or if we run into issues.

This is technically breaking for users of the new architecture who are relying on `dynamic_cast`, but that should be relatively rare. We could also maybe scope this so that the framework builds with RTTI, but user modules do not (so user's could `dynamic_cast` their own types).

Changelog:
[Android][Breaking] - Disable RTTI

Differential Revision: D45771601

fbshipit-source-id: dd574f3fa6eac1798edcb1b5c3b059e06cbe3446